feat: unify convert_* functions across gpu*.rs files #31104
